@@ -1,19 +1,18 @@
 {
-  "name": "client",
+  "name": "hex-2048",
   "version": "0.1.0",
-  "homepage": "/hexagonals2048/packages/client/build",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "classnames": "^2.2.6",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "classnames": "^2.3.2",
     "concurrently": "^7.6.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-hexgrid": "^1.0.3",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hexgrid": "^2.0.1",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "concurrently \"react-scripts start\" \"cd server && node src/index.js\"",
